separate script for cleaning
